Rename `utils.py` to `prompt.py`
